searobotics_surveyor/requirements/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# searobotics_surveyor\helpers\surveyor_messages_helper.py: 14
# searobotics_surveyor\surveyor.py: 5
geopy==2.4.1
# searobotics_surveyor\surveyor.py: 11
h5py==3.10.0
# searobotics_surveyor\clients\lidar_client.py: 46
# searobotics_surveyor\servers\lidar_server.py: 2
matplotlib==3.8.3
# searobotics_surveyor\clients\lidar_client.py: 47
# searobotics_surveyor\servers\lidar_server.py: 1
# searobotics_surveyor\surveyor.py: 10
numpy==1.26.3
# searobotics_surveyor\debugging_files\local_exo2_serial_reader.py: 6
# searobotics_surveyor\helpers\read_save_helper.py: 5
# searobotics_surveyor\helpers\waypoint_helper.py: 1
pandas==2.1.4
# searobotics_surveyor\servers\_camera_server.py: 7
# searobotics_surveyor\servers\camera_server.py: 2
pillow==10.2.0
# searobotics_surveyor\helpers\surveyor_messages_helper.py: 15
pynmea2==1.19.0
# searobotics_surveyor\debugging_files\local_exo2_serial_reader.py: 7
# searobotics_surveyor\debugging_files\socket_client.py: 1
# searobotics_surveyor\debugging_files\socket_server.py: 1
# searobotics_surveyor\servers\exo2_server.py: 4
pyserial==3.5
# searobotics_surveyor\clients\exo2_client.py: 2
# searobotics_surveyor\clients\lidar_client.py: 3
requests==2.31.0

# WARNING(pigar): some manual fixes might be required as pigar has detected duplicate requirements for the same import name (possibly for different submodules).
# WARNING(pigar): the following duplicate requirements are for the import name: cv2
# searobotics_surveyor\clients\camera_client.py: 1
# searobotics_surveyor\servers\_camera_server.py: 3
# searobotics_surveyor\servers\camera_server.py: 3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
